requirements.txt
sqlalchemy>=2.0
pydantic>=2.0
fastapi>=0.110
uvicorn[standard]>=0.29
fpdf2>=2.7
openpyxl>=3.1
cryptography>=41.0
gspread>=6.0
matplotlib>=3.8
pyyaml>=6.0
requests>=2.31
python-dotenv>=1.0
PySide6>=6.6
argon2-cffi>=23.1
rarfile>=4.1
pytest>=8.0
